requirements.txt
fastapi
uvicorn
sqlmodel
httpx
aiogram
python-dotenv
serpapi
psycopg2-binary  # PostgreSQL için bu şart!